[FIX] account: Fix various issues with round globally
--------------------------------------------------------
Fix the "same_tax_base" on tax totals.
--------------------------------------------------------

Suppose 2 lines:
- 100 with 13%
- 100 with 23%
=> The base are not displayed when printing the PDF because 'same_tax_base' is True.
However, the untaxed amount is 200 and both taxes have a base of 100.

--------------------------------------------------------
Fix the base amount when there is no tax.
--------------------------------------------------------

Suppose 2 lines having quantity=12.12, price_unit=12.12
=> The expected base amount is round(12.12 * 12.12 * 2) ~= 293.79
However, the current dispatching of taxes is only done when there is a tax set.

--------------------------------------------------------
Automatic fallback of the currency.
--------------------------------------------------------

When converting the record to a base line, automatically compute the currency_id
by looking to a company_currency_id or company_id field.

--------------------------------------------------------
Fix the untaxed amount in case of cash rounding line.
--------------------------------------------------------

Suppose a line of 15.68, cash rounding 0.05 half-up.
The untaxed amount returned by the tax totals was 16.70
with a cash rounding amount of 0.02.
However, 16.70 + 0.02 = 16.72... it's confusing for the user.
Let's subtract the cash rounding amount from the untaxed amount to
get 16.68 + 0.02 = 16.70.

--------------------------------------------------------
Start the test suite for the Portuguese certification.
--------------------------------------------------------

To use Odoo in Portugal, we have to get a certification that is mainly based on
the results given by our taxes computation.
This is currently a work in progress but this commit contains the beginning of
the test suite to get this certification.

closes odoo/odoo#191402

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>